+++ src/core/standard/index.ts
@@ -1,4 +1,13 @@
+import type { AlsoPromise } from '../../typings/extenders.js';
 import type { AnyFunction, Reverse } from '../../typings/helpers.js';
+
+export function execute(
+	condition: boolean,
+	correct: () => AlsoPromise<void> | AnyFunction<[]>,
+	otherwise: () => AlsoPromise<void> | AnyFunction<[]> = () => {}
+) {
+	condition ? correct() : otherwise();
+}
 
 /** identical - checks whether x and y have the same values */
 export function identical(x: unknown, y: unknown): boolean {
@@ -51,4 +60,8 @@ export function inverse<Function extends AnyFunction>(fn: Function) {
  */
 export function regexp(pattern: string, flags?: string): RegExp {
 	return new RegExp(pattern.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'), flags);
+}
+
+export function scope<T>(fn: () => T) {
+	return fn();
 }