+++ src/std/ntv/object.ts
@@ -10,6 +10,12 @@ export function clone<T>(i: T): T {
 	return iterate(i) as T;
 }
 
+export function create<T extends IndexSignature>(array: T[], i: any = null) {
+	const object = {} as { [K in T]: typeof i };
+	for (const key of array) object[key] = i;
+	return object;
+}
+
 export function entries<T extends object>(o: T) {
 	return Object.entries(o) as Entries<T>;
 }
@@ -38,13 +44,13 @@ export function iterate<T extends object, I = T[keyof T]>(
 	for (let i = 0; i < pairs.length; i++) {
 		const res = callback(pairs[i], i);
 		if (!res || res.length !== 2) continue;
-		memo.push(res as typeof memo[number]);
+		memo.push(res as (typeof memo)[number]);
 	}
 	return Object.fromEntries(memo) as any;
 }
 
 export function keys<T extends object>(o: T) {
-	return Object.keys(o) as Array<keyof T>;
+	return Object.keys(o) as Array<string & keyof T>;
 }
 
 export function pick<Keys extends readonly string[]>(keys: Narrow<Keys>) {